Organize file structure, create components
