+++ src/containers/App.css
@@ -7,10 +7,8 @@
 }
 
 .App-header {
-    background-color: #222;
     height: 150px;
     padding: 20px;
-    color: white;
 }
 
 .App-title {
